+++ R/deprecated.R
@@ -23,16 +23,6 @@ conf_int <- function(x, level = 0.95, type = "percentile",
   )
 }
 
-#' @rdname deprecated
-#' @export
-get_ci <- function(x, level = 0.95, type = "percentile",
-  point_estimate = NULL) {
-  .Deprecated("get_confidence_interval")
-  get_confidence_interval(
-    x, level = level, type = type, point_estimate = point_estimate
-  )
-}
-
 
 #' @rdname deprecated
 #' @export
@@ -41,9 +31,3 @@ p_value <- function(x, obs_stat, direction) {
   get_p_value(x = x, obs_stat = obs_stat, direction = direction)
 }
 
-#' @rdname deprecated
-#' @export
-get_pvalue <- function(x, obs_stat, direction) {
-  .Deprecated("get_p_value")
-  get_p_value(x = x, obs_stat = obs_stat, direction = direction)
-}